requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
biopython==1.87
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
faiss-cpu==1.13.2
filelock==3.29.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.1
pandas==3.0.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
RapidFuzz==3.14.5
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.6.0
typer==0.24.2
typing_extensions==4.15.0
tzdata==2026.1
